fix: extract workspace binding snapshot boundary
